Tidy-up: warning removal, and use int64 rather than __int64 or long long int.

+++ AdventOfCode2020/day03/day03.cpp
@@ -5,9 +5,9 @@
 
 std::vector<std::string> trees;
 
-int countTrees( int dx, int dy )
+int64 countTrees( int dx, int dy )
 {
-	int total = 0;
+	int64 total = 0;
 	int x = 0;
 	for ( int y = 0; y < trees.size(); y += dy, x += dx )
 	{
@@ -16,7 +16,7 @@ int countTrees( int dx, int dy )
 		if ( p == '#' )
 			++total;
 	}
-	printf( "Right %d, down %d: %d trees\n", dx, dy, total );
+	printf( "Right %d, down %d: %lld trees\n", dx, dy, total );
 	return total;
 }
 
@@ -38,13 +38,13 @@ int main()
 	fclose( fp );
 	delete[] buffer;
 
-	long long int total1 = countTrees( 1, 1 );
-	long long int total2 = countTrees( 3, 1 );
-	long long int total3 = countTrees( 5, 1 );
-	long long int total4 = countTrees( 7, 1 );
-	long long int total5 = countTrees( 1, 2 );
+	int64 total1 = countTrees( 1, 1 );
+	int64 total2 = countTrees( 3, 1 );
+	int64 total3 = countTrees( 5, 1 );
+	int64 total4 = countTrees( 7, 1 );
+	int64 total5 = countTrees( 1, 2 );
 
-	printf( "1: %d trees encountered\n", total2 );
+	printf( "1: %lld trees encountered\n", total2 );
 	printf( "2: %lld", total1*total2*total3*total4*total5 );
 
 	return 0;
